Create ACM Holding Animals
The total available floor space on the ark would have been over 100,000 square feet, which would be more floor space than in 20 standard sized
basketball courts. But in our story, we assume that the land dwelling air breathing animals is so many that the Ark can't contain all of them.

We know that the size of each kind of animal is different and the God has his own favorite. The God assigns each kind of animal a point to
show his favorite, and then lists points of every kind of animal to Noah. Noah must let some kinds of animals into the Ark and refuse
the others to maximize the total points. Noah cries for he doesn't know computer programming.

It 's an easy problem for you,right? So,rescue Noah!