Create ACM the legend of valiant emigration
The holy topic of human being,love,appears now.Firstly,I'd like to introduce a love song cited from <>
to you as follows.

The Woman
Your lips cover me with kisses;
your love is better than wine.
There is a fragrance about you;
the sound of your name recalls it.
No woman could help loving you.
Take me with you,and we'll run away.
be my king and take me to your room.
We will be happy together,
drink deep,and lose ourselves in love.
No wonder all women love you.
Women of Jerusalem,I am dark but beautiful,
dark as the desert tents of Kedar,
but beautiful as the curtains in Soloman's palace.
Don't look down on me because of my colour,
because the sun has tanned me.
My brotherw were angry with me
and made me work in the vineyard.
I had no time to care for myself.
Tell me,my love,
Where will you lead your flock to graze?
Where will they rest from the noonday sun?
Why should I need to look for you
among the flocks of the other shepherds?

The Man
Don't you know the place,loveliest of women?
Go and follow the flock;
find pasture for your goats near the tents of the shepherds.
You,my love,excite men as a mare excites the stallions of
Pharaoh's chariots.
Your hair is beautiful upon your cheeks
and falls along your neck like jewels.
But we will make for you a chain of gold with ornaments of silver.

The Woman
My king was lying on his couch,
and my perfume filled the air with fragrance.
My lover has the scent of myrrh as he lies upon my breasts.
My lover is like the wild flowers
that bloom in the vineyards at Engedi.

The Man
How beautiful you are,my love;
how your eyes shine with love!

The Woman
How handsome you are,my dearest;
how you delight me!
The green grass will be our bed;
the cedars will be the beams of our house,
and the cypress-trees the ceiling.
I am only a wild flower in Sharon,
a lily in a mountain valley.

The Man
Like a lily among thorns
is my darling among women.

The Woman
Like an apple-tree among the trees of the forest,
so is my dearest compared with other men.
I love to sit in its shadow,
and its fruit is sweet to my taste.
He brought me to his banqueting hall
and raised the banner of love over me.
Restore my strength with raisins
and refresh me with apples!
I am weak from passion.
His left hand is under my head,
and his right hand caresses me.
Promise me,women of Jerusalem;
swear by the swift deer and the gazelles
that you will not interrupt our love.

How beautiful the song is! The God is moved by the holy love between them and decides to help them run away to that place
having full-blown the green grass,the cedars,the cypress-trees and the flowers .There are many cities at that time,such as
Jerusalem,Bethlehem,Carmel,Hebron,Ziklog,Beersheba,Gath,Libnah,Ekron,Beth Horon,Shiloh,Gilgal,and so on.The two lovers
sets out from Jerusalem,and the destination is The Garden of Eden.There are roads between cities,each road has a
guard who has "pguard" power to beat and "sguard" speed to run.The two lovers can run away together,they have "plovers"
power to beat and "slovers" speed to run.  They can go through the road if and only if their power "plovers" is bigger than the guard and the speed "slovers" 

is bigger than the guard..The God help them calculate the cost of each road,and place a

heart-shaped elixir,associated with a letter to distinguish them,in each road.Your task is to minimize the cost

and display the letters of elixirs they will get in the way.