Create ACM Doraemon's flashlight
Doraemon, a robot cat from the 21st century, has a lot of magic tools. And he is always helping others. Little Ken, a little kid in Wuhan University,
is frequently teased by his classmates. So, Little Ken wants to be a big Ken, a strong Ken, instead of little Ken.
One day Doraemon heard of that the cherry blossoms in Wuhan University are more beautiful than those in Japan, so Doraemon flied to China by
his aircraft. While enjoying the blossoms, he happened to hear of Little Ken?s rough life. The kindhearted Doraemon decided to help Little Ken, you
know,to be a big Ken.

Doraemon picks out a flashlight from his pocket. The flashlight is one of his magic tools. All the objects illuminated by this flashlight will
become much bigger than before. So Doraemon points the flashlight to Little Ken. At the moment that Doraemon wants to push the button, Little
Ken cries, ?Wait a second!? Because he doubts this magic tool, Little Ken wants to test it first. He picks up a cubic and illuminates it.
The cubic changes its shape, as the following figure shows:

Little Ken realizes that the flashlight is actually processing a linear transform in the 3D Euclid Space. That means, given a 3*3 matrix A,
for example,

If a point?s original coordinate is, for instance, (1, 2, 3), then after the illuminated by the flashlight, the point moves to (4, 4, 2). That
is because the transform obeys the rule of a matrix multiplies a vector, as follows:

Now, given a matrix A, which represents the transform by the flashlight, your mission is to calculate the volume of the object that transformed
from a unit cubic.
The coordinates of the eight vertices of the unit cubic are: (0, 0, 0),(0, 1, 0),(1, 1, 0),(1, 0, 0),(0, 0, 1),(0, 1, 1),
(1, 1, 1),(1, 0, 1).
If the new object?s eight vertices become (0, 0, 0),(1, 0, -1),(1, 2, -1),(0, 2, 0),(1, 0, 1),(2, 0 , 0),(2, 2, 0),(1, 2, 1).
And the volume of the new object is 4.