Create ACM cherry blossom
March is wonderful in Wuhan University for the blooming cherry blossoms. Walking in the campus, you can smell the fragrance and feel the
romance with the tardiness falling of cherry blossom petals, and after a while you will see the ground will be covered by the beautiful
blossom petals.
 
Now, here comes the problem: the figure shown below is the cherry blossom petals on the ground, where the black points stand for blossom
petals. And this figure on the left is up-down symmetric as it is possible to cut the figure into two identical halves by the dashed line
shown in it, but the figure on the right is not.

This figure shows the sample input data

Given such a figure, can you tell me whether it is up-down symmetric or not. All the black points are different from each other.