Create ACM game on checkerboard
Suppose that you are given an  checkerboard and a checker. Let?s have a game on the checkerboard right now. Here, you should move the checker from the first row of the board to the last row of the board according to the following rule. At each step you may move the checker to one of three squares:

the square immediately below.
the square that is one below and one to the left(but only if the checker is not already in the leftmost column).
the square that is one below and one to the right(but only if the checker is not already in the rightmost column).
Consider each square of the checkerboard is placed p(i, j) dollars, where i is the row number and j is the column number of the square. You can move the checker from any square of the first row to any square of the last row, and pick up the dollars placed in the square you have passed. What is the maximum number of dollars you can gather along the way?