Create ACM find the max norm
As you know, matrix theory is widely used in the engineering and other fields. But do you know boolean matrix? It's a most useful matrix
in the computer science. In graph theory, IC design and compiler theory, boolean matrix plays an important role.
One day Little Dragonfly has to do his homework. But one Hard problem puzzles him. Now he wants you to help him to resolve this Hard
problem. The description of the Hard problem:
Note that Boolean matrix is the matrix that each entry in it is 0 or 1. We define the NORM of the boolean matrix as follow:
NORM = the sum of all the entry in the matrix.

And define two manipulations of the matrix:
A(n) : reverse every entry of the nth row;
B(n) : reverse every entry of the nth column;
Given a boolean matrix, you can get some matrices after a series of the two manipulations. Now you have to find the Max NORM of these
matrices.

Examples:
(1) Give a 22 matrix:
0 1
1 0
The Max NORM is 4.
Because we can get a matrix
1 1
1 1
after A(1),B(2). Then the NORM is the largest.
(2) Give another 22 matrix
0 1
1 1
The Max NORM is 3.