Create ACM cut the apple
On the way home, DragonFly has got an extraordinarily large apple. He would like to share it with his friends. Now he wants to get
as many pieces as possible with n cuts. But he cannot decide how to cut it, because the apple is rather delicious. Now he needs your help.