Create ACM BG
Magicpig and Acm like to compete with each other recently, and the loser has to "BG" the winner, that means, if Magicpig wins, Acm will treat him to a meal. But Acm is so powerful that Magicpig has to "BG" him almost every time. Everybody is very interested in this. So more and more ACMers are attracted to participate in this game.

One day, Peipei said: "It is unfair that one person should 'BG' all the others. I suggest that 50% of the participants should pay the bill. That means, if there are n participants, then Ceil(n/2) person should 'BG' the other n-Ceil(n/2) person." This is a good idea! So all of us agree to do so. This is the story of "BG".

Now you task is: given a number n -- the total number of participants, you have to calculate how many people should pay the bill.