Create ACM crazy game
What a hot day! The summer of Wuhan is always hot. One day, Flymouse and his classmates went to the natatorium together. The natatorium developed an amusing team game. They built a “bridge” in the center of the pool. The bridge only could hold two students. Strangely, it could be changed 180 degrees. By this way two students can change their orders when they are neighbors. Now, everyone of N students is given an unique number between 1 and N. After several changes, they form a queue by the orders:1  2  3  4 …… N.Flymouse is not clever enough and cannot solve this problem by the minimum times of changes. So he asks for your help.