Create ACM what is left
the same number. When all the books are marked, it is easy to find how many are the same. :-)
Now QingQing has a list of books( given by numbers they are marked). In order to buy a gift for her dear brother, she plans to sell
some books. But she wants to keep one copy for each kind. After selling, what books is left?
Please write a program to tell QingQing the answer.