Create ACM 2D path problem
Imagine that we draw a straight line that connects the last point to the origin point in the path. (The line that is drawn in dots in the figure above) In the figure above, we can see each segment has its own direction in the drawing process. And the figure is divided into several parts. Now, we want to know the total area of all the parts. But, each part's area is a little different from common. In each part, if its boundary is clockwise, the area of this part is negative, otherwise, its area is positive.
Now, given the drawing path, can you tell me the total area described above'