Create ACM zap
Dongfangxu and zap are two good friends. They have a perverse hobby, which is playing string. In a series of messy string to find their own name appears the most times in a row, who appears more times, then stick a note on another person's face. They need a referee to help them, and they have an agreement. If two people's names appear as many times in succession, then dongfangxu wins. The referee looked at the string, my God. It's a mess. It's hard to count, so he asks you for help.