Create ACM forest
Eight acmers graduated in June, and they were:

"littleken","knuthocean","dongfangxu","zap","kittig","robertcui","forest","flirly"。

They decided to corrupt after graduation, but no one wanted to pay BG, so forest thought of a way to get the computer to generate a string of their names, and see

Who has the most names in the string and whose names appear the most, then he will have BG others, but because the string is long, they are lazy to count,

Give you this task. Who calls you their younger brother! As acmer, you make a program to solve it.