Create ACM bullseye
A simple dartboard consists of a flat, circular piece of cork with concentric rings drawn on it. Darts are thrown at the board by players in
an attempt to hit the center of the dartboard (the Bullseye). The region between each pair of rings (or the center and the first ring)
represents a certain point value. The closer the region is to the center of the dartboard, the more points the region is worth, as shown in
the diagram below:



Ring radii are at 3", 6", 9", 12" and 15" (the Bullseye has a diameter of 6"). A game of Simple Darts between two players is played as
follows. The first player throws 3 darts at the board. A score is computed by adding up the point values of each region that a dart lands in.
The darts are removed. The second player throws 3 darts at the board; the score for player two is computed the same way as it is for player
one. The player with the higher score wins.

For this problem, you are to write a program that computes the scores for two players, and determine who, if anyone, wins the game. If a dart
lands exactly on a ring (region boundary), the higher point value is awarded. Any dart outside the outer ring receives no points. For the
purposes of this problem, you can assume that a dart has an infinitely fine point and can not land paritially on a ring; it is either on the
ring or it is not on the ring. Standard double precision floating point operations will be should be used.