Create ACM Excel-lent problem
A certain spreadsheet program labels the columns of a spreadsheet using letters. Column 1 is labeled as A", column 2 asB", ..., column 26
as Z". When the number of columns is greater than 26, another letter is used. For example, column 27 isAA", column 28 is AB" and column <br/>52 isAZ". It follows that column 53 would be BA" and so on. Similarly, when columnZZ" is reached, the next column would be AAA", <br/>thenAAB" and so on.

The rows in the spreadsheet are labeled using the row number. Rows start at 1.

The designation for a particular cell within the spreadsheet is created by combining the column label with the row label. For example, the
upper-left most cell would be A1". The cell at column 55 row 23 would beBC23".

You will write a program that converts numeric row and column values into the spreadsheet designation.