Create ACM model rocket height
One method for determining the height achieved by a model rocket is to have three observers A, B and C equally spaced D feet apart along a
line at one edge of the flat test field. Each observer has a theodolite or some other device for measuring angle above the horizontal of a
distant object. Each measuring device is on a stand H feet above the field. When a rocket is fired, near the top of its flight, it deploys a
parachute and emits a puff of dust. Each observer measures the angle above the horizontal of the puff of dust from their location. From these
angles α for A, β for B and γ for C, the height of the rocket above the field can be determined.

This problem is to write a program which, given the parameters D (the distance between observers along the line in feet), H (the distance of
the measuring device above the field in feet), α (the angle of the rocket above the horizontal in degrees measured by the leftmost observer
A), β (the angle above the horizontal in degrees observed by the center observer B) and γ (the angle above the horizontal in degrees
measured by the rightmost observer C), computes the height of the rocket above the field in feet to the nearest foot.