Create ACM Olympic
"One World One Dream" fully reflects the essence and the universal values of the Olympic spirit ---- Unity, Friendship, Progress, Harmony,
Participation and Dream. It expresses the common wishes of people all over the world, inspired by the Olympic ideals, to strive for a bright future
of Mankind. In spite of the differences in colors, languages and races, we share the charm and joy of the Olympic Games, and together we seek for the ideal of Mankind for peace. We belong to the same world and we share the same aspirations and dreams.

"One World One Dream" is a profound manifestation of the core concepts of the Beijing Olympic Games. It reflects the values of harmony connoted in the concept of "People's Olympics", the core and soul of the three concepts -- "Green Olympics, High-tech Olympics and People's Olympics".
While "Harmony of Man with Nature" and "Peace Enjoys Priority" are the philosophies and ideals of the Chinese people since ancient times in their pursuit of the harmony between Man and Nature and the harmony among people, building up a harmonious society and achieving harmonious
development are the dream and aspirations of ours. It is our belief that peace and progress, harmonious development, living in amity, cooperation and mutual benefit, and enjoying a happy life are the common ideals of the people throughout the world.

"One World, One Dream" is simple in expressions, but profound in meaning. It is of China, and also of the world. It conveys the lofty ideal of the
people in Beijing as well as in China to share the global community and civilization and to create a bright future hand in hand with the people from the rest of the world. It expresses the firm belief of a great nation, with a long history of 5,000 years and on its way towards modernization, that
is committed to peaceful development, harmonious society and people's happiness. It voices the aspirations of 1.3 billion Chinese people to
contribute to the establishment of a peaceful and bright world.

The English translation of the slogan is distinctive in sentence structure. The two "One"s are perfectly used in parallel, and the words "World" and "Dream" form a good match. The slogan is simple, meaningful, inspiring, and easy to remember, read and spread.

In Chinese, the word "tongyi", which means "the same", is used for the English word "One". It highlights the theme of "the whole Mankind lives in the same world and seeks for the same dream and ideal".

Please help us write a program to count down days to the upcoming Beijing 2008 Olympic games.