Create ACM catch that cow
Farmer John has been informed of the location of a fugitive cow and
wants to catch her immediately. He starts at a point N (0 <= N <=
100,000) on a number line and the cow is at a point K (0 <= K <=
100,000) on the same number line. Farmer John has two modes of
transportation: walking and teleporting.

* Walking: FJ can move from any point X to the points X-1 or
X+1 in a single minute

* Teleporting: FJ can move from any point X to the point 2*X
in a single minute.

If the cow, unaware of its pursuit, does not move at all, how long
does it take for Farmer John to retrieve it?