Create ACM Negative weight
When we write a decimal book, its value can be expressed by multiplying each digit by a power of 10. For example:

123 = 1×10^2 + 2×10^1 + 3×10^0

Generally speaking, for r-ary number n, its absolute value can be multiplied by the number of each digit by the power of R

N = a[n]×R^(n) + a[n-1]×R^(n-1) + ... + a[0]×R^(0)

To express. Here R can be positive or negative. When R is negative, we call it negative weight. No matter whether R is positive or negative, we use {0,1,..., | R | - 1} to represent the bits of R-base number. If | R | > 10, we will also use capital letters for numbers. For example, for a hexadecimal number, a stands for 10 (decimal), B stands for 11 (decimal), and f stands for 15 (decimal).

One of the advantages of using negative weights is that we don't need to use the minus sign? -?. For example, the decimal number - 15 is represented by - 2, which is 110001:

-15 = 1×(-2)^5 + 1×(-2)^4 + 0×(-2)^3 + 0×(-2)^2 + 0×(-2)^1 + 1×(-2)^0